fix: export cli flags interfaces
