+++ src/index.ts
@@ -1,6 +1,3 @@
-import build from './build';
-import { BuildCliFlags } from './commands/build.parser';
-import { PreviewCliFlags } from './commands/preview.parser';
-import preview from './preview';
-
-export { build, preview, BuildCliFlags, PreviewCliFlags };
+export { build, BuildCliFlags } from './build';
+export { init, InitCliFlags } from './init';
+export { preview, PreviewCliFlags } from './preview';
